+++ core/src/main/java/org/polypheny/db/adapter/DataStore.java
@@ -51,7 +51,7 @@ public abstract class DataStore extends Adapter implements ExtensionPoint {
     protected final transient Catalog catalog = Catalog.getInstance();
 
 
-    public DataStore( final int adapterId, final String uniqueName, final Map<String, String> settings, final boolean persistent ) {
+    public DataStore( final long adapterId, final String uniqueName, final Map<String, String> settings, final boolean persistent ) {
         super( adapterId, uniqueName, settings );
         this.persistent = persistent;
